requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
bcrypt==3.1.4
betfairlightweight==1.5.13
bleach==2.1.2
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
ciso8601==1.0.7
cryptography==2.3.1
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
future==0.16.0
html5lib==1.0.1
idna==2.7
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
iso8601==0.1.12
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
mysql-connector-python==8.0.12
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.0
pandas==0.22.0
pandas-datareader==0.6.0
pandocfilters==1.4.2
paramiko==2.4.0
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.5.2
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
pyparsing==2.2.0
pyTelegramBotAPI==3.6.6
python-dateutil==2.6.1
python-telegram-bot==11.1.0
pytz==2018.3
PyYAML==3.13
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
requests-file==1.4.3
requests-ftp==0.3.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.5.0
serial==0.0.66
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
ujson==1.35
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
wrapt==1.10.11
